backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
sentencepiece==0.1.99
accelerate==0.24.0
nltk==3.8.1
textstat==0.7.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3